requirements.txt
matplotlib==1.4.3
mock==1.0.1
nose==1.3.7
numpy==1.9.2
Pillow==2.8.2
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
scipy==0.15.1
six==1.9.0
wheel==0.24.0
